Improved MediaController. Now can set different maxUploadSize for extensions
