+++ src/OnlineSales/appsettings.json
@@ -14,7 +14,7 @@
     "Server": "localhost",
     "Port": "5432",
     "UserName": "postgres",
-    "Password": "postgres",
+    "Password": "147852",
     "Database": "onlinesales"
   },
   "Elastic": {
@@ -61,7 +61,7 @@
       "BatchSize": "5000"
     }
   },
-  "Images": {
+  "Media": {
     "Extensions": [
       ".png",
       ".jpg",
@@ -69,9 +69,19 @@
       ".ico",
       ".webp",
       ".svg",
-      ".gif"
+      ".gif",
+      ".mp4"
     ],
-    "MaxSize": "500KB"
+    "MaxSize": [
+      {
+        "Extension": "default",
+        "MaxSize": "500KB"
+      },
+      {
+        "Extension": ".mp4",
+        "MaxSize": "3000KB"
+      }
+    ]
   },
   "CacheProfiles": [
     {
